Turn off all output processing
It translates each newline ("\n") we print into a carriage return followed by a newline ("\r\n").
The terminal requires both of these characters in order to start a new line of text.
The carriage return moves the cursor back to the beginning of the current line, and the newline moves the cursor down a line, scrolling the screen if necessary.
These two distinct operations originated in the days of typewriters and teletypes.