key press unblocked
bufio.Reader.ReadByte() blocks the main process, so nothing else can run without further key press.
With VMIN and VTIME you can set the terminal input to non-blocking, but this option is only available for Linux and macOS, not for Windows.
For cross-platform compatibility, the key press process is outsourced to a goroutine.