OPOST is not disabled
`terminalRawConfigure` is called via stdin, but OPOST belongs to stdout.
The terminal device is the same, but the stdin and stdout handles are different.
Therefore, `terminalRawConfigure` must be called with both stdin and stdout handles.
The editor only manipulates one terminal structure, so `terminalRawConfigure` does not need to be split into stdin and stdout.